# server/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API | Baú da Saúde",
    "version": "1.0.0",
    "description": "Documentação da API Baú da Saúde"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/users": {
      "post": {
        "summary": "Cria um novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso"
          },
          "400": {
            "description": "Erro na criação do usuário"
          }
        }
      },
      "get": {
        "summary": "Obtém todos os usuários",
        "responses": {
          "200": {
            "description": "Lista de usuários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro no servidor"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Obtém um usuário pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "500": {
            "description": "Erro no servidor"
          }
        }
      }
    },
    "/users/email/{email}": {
      "get": {
        "summary": "Obtém um usuário pelo email",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "500": {
            "description": "Erro no servidor"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Login de um usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso"
          },
          "401": {
            "description": "Credenciais inválidas"
          },
          "500": {
            "description": "Erro no servidor"
          }
        }
      }
    },
    "/forgot-password": {
      "post": {
        "summary": "Solicita redefinição de senha",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPassword"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email de redefinição enviado"
          },
          "400": {
            "description": "Usuário não encontrado"
          },
          "500": {
            "description": "Erro no servidor"
          }
        }
      }
    },
    "/reset-password/{token}": {
      "post": {
        "summary": "Redefine a senha do usuário",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPassword"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Senha redefinida com sucesso"
          },
          "400": {
            "description": "Token inválido ou expirado"
          },
          "500": {
            "description": "Erro no servidor"
          }
        }
      }
    },
    "/delete-user": {
      "delete": {
        "summary": "Deleta um usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário deletado com sucesso"
          },
          "400": {
            "description": "Email ou ID é necessário"
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "500": {
            "description": "Erro no servidor"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Login": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "ForgotPassword": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "ResetPassword": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        }
      },
      "DeleteUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        }
      }
    }
  }
}